pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "steambot"
version = "1.2.0"
description = "A Discord bot for Steam news feeds."
requires-python = ">=3.11"
readme = "README.md"
license-files = { paths = ["LICENSE"] }
keywords = ["Discord", "bot", "Steam", "news", "RSS"]
dependencies = [
  "beautifulsoup4~=4.11",
  "py-cord~=2.2",
  "python_dateutil~=2.8",
  "requests~=2.28",
  "Whoosh~=2.7"
]

[project.urls]
"Source code" = "https://github.com/Faxmachinen/steamnews"

[project.scripts]
cli-name = "steambot.scripts:main"
